move clients to their own folders
client is not the perfect name but is
the best i got, can be renamed later.
